Move privacy to app/models/table #2781
Also move it to CartoDB::Table namespace
